PawPal-Review/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
gunicorn==21.2.0
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.1